changes to follow button and state
